[add] get ranking from firebase and dispatch this data
