+++ app/src/main/java/voicebattle/com/shaya/voicebattle/ranking/RankingEntity.kt
@@ -2,5 +2,5 @@ package voicebattle.com.shaya.voicebattle.ranking
 
 class RankingEntity(
         var name:String,
-        var power:Int) {
+        var power:Long) {
 }
